Move MemWAL HNSW into lance crate
