--- rust/lance-hnsw/Cargo.toml
@@ -1,34 +1,0 @@
-[package]
-name = "lance-hnsw"
-version.workspace = true
-edition.workspace = true
-authors.workspace = true
-license.workspace = true
-repository.workspace = true
-readme.workspace = true
-description = "Fast HNSW graph primitives for Lance"
-keywords.workspace = true
-categories.workspace = true
-rust-version.workspace = true
-
-[dependencies]
-arc-swap.workspace = true
-arrow-array.workspace = true
-arrow-schema.workspace = true
-crossbeam-queue.workspace = true
-lance-core.workspace = true
-lance-linalg.workspace = true
-rand.workspace = true
-rayon.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-
-[[bin]]
-name = "lance_hnsw_bench"
-path = "src/bin/lance_hnsw_bench.rs"
-
-[dev-dependencies]
-lance-index.workspace = true
-
-[lints]
-workspace = true
