+++ rust/lance/Cargo.toml
@@ -19,6 +19,7 @@ no-default-features = true
 ignored = ["tracing-chrome"]
 
 [dependencies]
+arc-swap = { workspace = true }
 lance-arrow = { workspace = true }
 lance-core = { workspace = true }
 lance-datafusion = { workspace = true }
@@ -26,7 +27,6 @@ lance-encoding = { workspace = true }
 lance-file = { workspace = true }
 lance-io = { workspace = true }
 lance-linalg = { workspace = true }
-lance-hnsw = { workspace = true }
 lance-index = { workspace = true }
 lance-namespace = { workspace = true }
 lance-tokenizer = { workspace = true }
@@ -46,6 +46,7 @@ byteorder.workspace = true
 bytes.workspace = true
 chrono.workspace = true
 clap = { version = "4.1.1", features = ["derive"], optional = true }
+crossbeam-queue = { workspace = true }
 crossbeam-skiplist.workspace = true
 # This is already used by datafusion
 dashmap = "6"
@@ -63,6 +64,7 @@ roaring.workspace = true
 tokio.workspace = true
 url.workspace = true
 rand.workspace = true
+rayon.workspace = true
 futures.workspace = true
 uuid.workspace = true
 arrow.workspace = true
@@ -224,6 +226,10 @@ harness = false
 
 [[bench]]
 name = "mem_wal_recall_hnsw"
+harness = false
+
+[[bench]]
+name = "mem_wal_hnsw_bench"
 harness = false
 
 [lints]
